+++ CHANGES.rst
@@ -18,6 +18,9 @@ Changelog
 2.6.0 (2024-04-02)
 ------------------
 
+- Manage the user's phone attribute in booking-schema
+  [mamico]
+
 - Return empty data from the @day endpoint if requested date is out of PrenotazioniFolder range
   [folix-01]